+++ src/tests.ts
@@ -26,8 +26,7 @@ export async function testLooseFiles(api: types.IExtensionApi): Promise<types.IT
       let iniContent = (await fs.readFileAsync(iniPath, 'utf-8')) ?? '';
       ini = parse(iniContent);
       return ini?.Archive?.bInvalidateOlderFiles === '1'
-        && ini?.Archive?.sResourceDataDirsFinal === ''
-        && (ini?.Display?.sPhotoModeFolder === undefined || ini?.Display?.sPhotoModeFolder !== 'Photos');
+        && ini?.Archive?.sResourceDataDirsFinal === '';
     } catch (err) {
       log('warn', `${archiveInvalidationTag} - INI not setup: ${iniPath}`);
       return false;
@@ -40,7 +39,7 @@ export async function testLooseFiles(api: types.IExtensionApi): Promise<types.IT
       const res: types.ITestResult = {
         description: {
           short: 'StarfieldCustom.ini not configured',
-          long: 'Similar to Fallout 4, Starfield requires certain INI tweaks to be set in order to properly load loose files (i.e. those not packed in BA2 archives). There are a lot of mods out there which provide instructions for users to add these tweaks to a "StarfieldCustom.ini" file in the "Documents\\My Games\\Starfield" folder. If Vortex detects that this ini doesn\'t exist or is incorrect, it will notify the user and ask to fix it. If fix is requested, it will add or adjust the "bInvalidateOlderFiles" and "sResourceDataDirsFinal" values without changing any other settings you might\'ve added manually. Additionally, Vortex will apply a tweak to re-route your Photo Mode captures to Data\\Textures\\Photos (unless you\'ve already set it to something else) and there is now a button inside Vortex to quickly open this folder.',
+          long: 'Similar to Fallout 4, Starfield requires certain INI tweaks to be set in order to properly load loose files (i.e. those not packed in BA2 archives). There are a lot of mods out there which provide instructions for users to add these tweaks to a "StarfieldCustom.ini" file in the "Documents\\My Games\\Starfield" folder. If Vortex detects that this ini doesn\'t exist or is incorrect, it will notify the user and ask to fix it. If fix is requested, it will add or adjust the "bInvalidateOlderFiles" and "sResourceDataDirsFinal" values without changing any other settings you might\'ve added manually.',
         },
         severity: 'warning',
         automaticFix: ((async () => {
@@ -64,9 +63,7 @@ export async function testLooseFiles(api: types.IExtensionApi): Promise<types.IT
             // Set required settiings on ini object and convert back to writeable string
             ini.Archive.bInvalidateOlderFiles = '1';
             ini.Archive.sResourceDataDirsFinal = '';
-            if (ini.Display?.sPhotoModeFolder !== undefined && ini.Display?.sPhotoModeFolder === 'Photos') {
-              delete ini.Display.sPhotoModeFolder;
-            }
+
             const newIniContent = sanitizeIni(stringify(ini, { retainComments: true, whitespace: true }));
             log('info', `${archiveInvalidationTag} - New INI: \n${newIniContent}`, ini);