+++ CHANGELOG.md
@@ -6,8 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [0.6.7] - 2024-03-19
 
-- Modified custom preferences ini test to ensure that the sPhotoModeFolder entry is removed unless manually set by the user.
-- Removed "Open Photos Folder" toolbar button as Vortex no longer manages the photo folder in any way.
+- Removed all logic pertaining to the sPhotoModeFolder ini entry (Please add/remove/modify the ini entry manually if needed)
 
 ## [0.6.6] - 2023-12-20